Move parser to own module
Then we can autoload it only if used. This makes room for the coming
Logfmt::Logger-ish stuff; one use case is the parser to read log lines,
and the other is the Logger to write them. We don't want to force one
use case to load the other into memory.